Don't depend on `guzzlehttp/psr7` just for testing
Otherwise, we would have to increase `guzzlehttp/psr7` (ddd7fa8) to support
recent PHP versions. This is a bit excessive for a library from which we only
need a single class exclusively for testing: The `UploadedFile` class was only
used to test our `FileValidator` and is now replaced by a minimal
implementation of `PSR-7#UploadedFileInterface`.